netmon/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.0
markdown==3.4.4
psutil==5.9.5
speedtest-cli==2.1.3